Move FiloContext to public value class so it's more accessible from Java
