+++ spark/src/main/scala/filodb.spark/package.scala
@@ -7,6 +7,7 @@ import org.apache.spark.sql.{SQLContext, SaveMode, DataFrame, Row}
 import org.apache.spark.sql.types.StructType
 import scala.concurrent.{Await, Future}
 import scala.concurrent.duration._
+import scala.language.implicitConversions
 import scala.language.postfixOps
 
 import filodb.core._
@@ -56,13 +57,15 @@ package object spark extends StrictLogging {
   import FiloSetup.metaStore
   import RowSource._
 
-  private def ingestRddRows(coordinatorActor: ActorRef,
-                            dataset: DatasetRef,
-                            columns: Seq[String],
-                            version: Int,
-                            rows: Iterator[Row],
-                            writeTimeout: FiniteDuration,
-                            partitionIndex: Int): Unit = {
+  val sparkLogger = logger
+
+  private[spark] def ingestRddRows(coordinatorActor: ActorRef,
+                                   dataset: DatasetRef,
+                                   columns: Seq[String],
+                                   version: Int,
+                                   rows: Iterator[Row],
+                                   writeTimeout: FiniteDuration,
+                                   partitionIndex: Int): Unit = {
     val props = RddRowSourceActor.props(rows, columns, dataset, version, coordinatorActor)
     val rddRowActor = FiloSetup.system.actorOf(props, s"${dataset}_${version}_$partitionIndex")
     actorAsk(rddRowActor, Start, writeTimeout) {
@@ -92,246 +95,95 @@ package object spark extends StrictLogging {
     }
   }
 
-  implicit class FiloContext(sqlContext: SQLContext) {
+  private[spark] def runCommands[B](cmds: Set[Future[Response]]): Unit = {
+    val responseSet = Await.result(Future.sequence(cmds), 5 seconds)
+    if (!responseSet.forall(_ == Success)) throw new RuntimeException(s"Some commands failed: $responseSet")
+  }
 
-    /**
-     * Creates a DataFrame from a FiloDB table.  Does no reading until a query is run, but it does
-     * read the schema for the table.
-     * @param dataset the name of the FiloDB dataset to read from
-     * @param database the database / Cassandra keyspace to read the dataset from
-     * @param version the version number to read from
-     * @param splitsPerNode the parallelism or number of splits per node
-     */
-    def filoDataset(dataset: String,
-                    database: Option[String] = None,
-                    version: Int = 0,
-                    splitsPerNode: Int = 4): DataFrame =
-      sqlContext.baseRelationToDataFrame(FiloRelation(DatasetRef(dataset, database), version, splitsPerNode)
-                                                     (sqlContext))
+  import filodb.spark.TypeConverters._
 
-    private def runCommands[B](cmds: Set[Future[Response]]): Unit = {
-      val responseSet = Await.result(Future.sequence(cmds), 5 seconds)
-      if (!responseSet.forall(_ == Success)) throw new RuntimeException(s"Some commands failed: $responseSet")
-    }
+  private[spark] def dfToFiloColumns(df: DataFrame): Seq[DataColumn] = dfToFiloColumns(df.schema)
 
-    import filodb.spark.TypeConverters._
-
-    private def dfToFiloColumns(df: DataFrame): Seq[DataColumn] = dfToFiloColumns(df.schema)
-
-    private def dfToFiloColumns(schema: StructType): Seq[DataColumn] = {
-      schema.map { f =>
-        DataColumn(0, f.name, "", -1, sqlTypeToColType(f.dataType))
-      }
-    }
-
-    private def checkAndAddColumns(dfColumns: Seq[DataColumn],
-                                   dataset: DatasetRef,
-                                   version: Int): Unit = {
-      // Pull out existing dataset schema
-      val schema = parse(metaStore.getSchema(dataset, version)) { schema =>
-        logger.info(s"Read schema for dataset $dataset = $schema")
-        schema
-      }
-
-      // Translate DF schema to columns, create new ones if needed
-      val dfSchema = dfColumns.map { col => col.name -> col }.toMap
-      val matchingCols = dfSchema.keySet.intersect(schema.keySet)
-      val missingCols = dfSchema.keySet -- schema.keySet
-      logger.info(s"Matching columns - $matchingCols\nMissing columns - $missingCols")
-
-      // Type-check matching columns
-      val matchingTypeErrs = matchingCols.collect {
-        case colName: String if dfSchema(colName).columnType != schema(colName).columnType =>
-          (colName, dfSchema(colName).columnType, schema(colName).columnType)
-      }
-      if (matchingTypeErrs.nonEmpty) throw ColumnTypeMismatch(matchingTypeErrs)
-
-      if (missingCols.nonEmpty) {
-        val addMissingCols = missingCols.map { colName =>
-          val newCol = dfSchema(colName).copy(dataset = dataset.dataset, version = version)
-          metaStore.newColumn(newCol, dataset)
-        }
-        runCommands(addMissingCols)
-      }
-    }
-
-    // Checks for schema errors via RichProjection.make, and returns created Dataset object
-    private def makeAndVerifyDataset(datasetRef: DatasetRef,
-                                     rowKeys: Seq[String],
-                                     segmentKey: String,
-                                     partitionKeys: Seq[String],
-                                     chunkSize: Option[Int],
-                                     dfColumns: Seq[Column]): Dataset = {
-      val options = Dataset.DefaultOptions
-      val options2 = chunkSize.map { newSize => options.copy(chunkSize = newSize) }.getOrElse(options)
-      val dataset = Dataset(datasetRef, rowKeys, segmentKey, partitionKeys).copy(options = options2)
-
-      // validate against schema.  Checks key names, computed columns, etc.
-      RichProjection.make(dataset, dfColumns).recover {
-        case err: RichProjection.BadSchema => throw BadSchemaError(err.toString)
-      }
-
-      dataset
-    }
-
-    // This doesn't create columns, because that's in checkAndAddColumns.
-    private def createNewDataset(dataset: Dataset): Unit = {
-      logger.info(s"Creating dataset ${dataset.name}...")
-      actorAsk(FiloSetup.coordinatorActor, CreateDataset(dataset, Nil)) {
-        case DatasetCreated =>
-          logger.info(s"Dataset ${dataset.name} created successfully...")
-        case DatasetError(errMsg) =>
-          throw new RuntimeException(s"Error creating dataset: $errMsg")
-      }
-    }
-
-    private def truncateDataset(dataset: Dataset, version: Int): Unit = {
-      logger.info(s"Truncating dataset ${dataset.name}")
-      actorAsk(FiloSetup.coordinatorActor,
-               TruncateProjection(dataset.projections.head, version), 1.minute) {
-        case ProjectionTruncated => logger.info(s"Truncation of ${dataset.name} finished")
-        case DatasetError(msg) => throw NotFoundError(s"$msg - (${dataset.name}, ${version})")
-      }
-    }
-
-    private def deleteDataset(dataset: DatasetRef): Unit = {
-      logger.info(s"Deleting dataset $dataset")
-      parse(FiloSetup.metaStore.deleteDataset(dataset)) { resp => resp }
-    }
-
-    /**
-     * Creates (or recreates) a FiloDB dataset with certain row, segment, partition keys.  Only creates the
-     * dataset/projection definition and persists the dataset metadata; does not actually create the column
-     * definitions (that is done by the insert step).  The exact behavior depends on the mode:
-     *   Append  - creates the dataset if it doesn't exist
-     *   Overwrite - creates the dataset, deleting the old definition first if needed
-     *   ErrorIfExists - throws an error if the dataset already exists
-     *
-     * For the other paramter definitions, please see saveAsFiloDataset().
-     */
-    private[spark] def createOrUpdateDataset(schema: StructType,
-                                             dataset: DatasetRef,
-                                             rowKeys: Seq[String],
-                                             segmentKey: String,
-                                             partitionKeys: Seq[String],
-                                             chunkSize: Option[Int] = None,
-                                             mode: SaveMode = SaveMode.Append): Unit = {
-      FiloSetup.init(sqlContext.sparkContext)
-      val partKeys = if (partitionKeys.nonEmpty) partitionKeys else Seq(Dataset.DefaultPartitionColumn)
-      val dfColumns = dfToFiloColumns(schema)
-
-      val datasetObj = try {
-        Some(getDatasetObj(dataset))
-      } catch {
-        case e: NotFoundError => None
-      }
-      (datasetObj, mode) match {
-        case (None, SaveMode.Append) | (None, SaveMode.Overwrite) | (None, SaveMode.ErrorIfExists) =>
-          val ds = makeAndVerifyDataset(dataset, rowKeys, segmentKey, partKeys, chunkSize, dfColumns)
-          createNewDataset(ds)
-        case (Some(dsObj), SaveMode.ErrorIfExists) =>
-          throw new RuntimeException(s"Dataset $dataset already exists!")
-        case (Some(dsObj), SaveMode.Overwrite) =>
-          val ds = makeAndVerifyDataset(dataset, rowKeys, segmentKey, partKeys, chunkSize, dfColumns)
-          deleteDataset(dataset)
-          createNewDataset(ds)
-        case (_, _) =>
-          logger.info(s"Dataset $dataset definition not changed")
-      }
-    }
-
-    /**
-     * Saves a DataFrame in a FiloDB Table
-     * - Creates columns in FiloDB from DF schema if needed
-     *
-     * @param df the DataFrame to write to FiloDB
-     * @param dataset the name of the FiloDB table/dataset to read from
-     * @param rowKeys the name of the column(s) used as the row primary key within each partition.
-     *                May be computed functions. Only used if mode is Overwrite and
-     * @param segmentKey the name of the column or computed function used to group rows into segments and
-     *                   to sort the partition by.
-     * @param partitionKeys column name(s) used for partition key.  If empty, then the default Dataset
-     *                      partition key of `:string /0` (a constant) will be used.
-     *
-     *          Partitioning columns could be created using an expression on another column
-     *          {{{
-     *            val newDF = df.withColumn("partition", df("someCol") % 100)
-     *          }}}
-     *          or even UDFs:
-     *          {{{
-     *            val idHash = sqlContext.udf.register("hashCode", (s: String) => s.hashCode())
-     *            val newDF = df.withColumn("partition", idHash(df("id")) % 100)
-     *          }}}
-     *
-     *          However, note that the above methods will lead to a physical column being created, so
-     *          use of computed columns is probably preferable.
-     *
-     * @param database the database/keyspace to write to, optional.  Default behavior depends on ColumnStore.
-     * @param mode the Spark SaveMode - ErrorIfExists, Append, Overwrite, Ignore
-     * @param options various IngestionOptions, such as timeouts, version to write to, etc.
-     */
-    def saveAsFilo(df: DataFrame,
-                   dataset: String,
-                   rowKeys: Seq[String],
-                   segmentKey: String,
-                   partitionKeys: Seq[String],
-                   database: Option[String] = None,
-                   mode: SaveMode = SaveMode.Append,
-                   options: IngestionOptions = IngestionOptions()): Unit = {
-      val IngestionOptions(version, chunkSize, writeTimeout, flushAfterInsert) = options
-      val ref = DatasetRef(dataset, database)
-      createOrUpdateDataset(df.schema, ref, rowKeys, segmentKey, partitionKeys, chunkSize, mode)
-      insertIntoFilo(df, dataset, version, mode == SaveMode.Overwrite,
-                     database, writeTimeout, flushAfterInsert)
-    }
-
-    /**
-     * Implements INSERT INTO into a Filo Dataset.  The dataset must already have been created.
-     * Will check and add any extra columns from the DataFrame into the dataset, but column type
-     * mismatches will result in an error.
-     * @param overwrite if true, first truncate the dataset before writing
-     */
-    def insertIntoFilo(df: DataFrame,
-                       datasetName: String,
-                       version: Int = 0,
-                       overwrite: Boolean = false,
-                       database: Option[String] = None,
-                       writeTimeout: FiniteDuration = DefaultWriteTimeout,
-                       flushAfterInsert: Boolean = true): Unit = {
-      val filoConfig = FiloSetup.initAndGetConfig(sqlContext.sparkContext)
-      val dfColumns = dfToFiloColumns(df)
-      val columnNames = dfColumns.map(_.name)
-      val dataset = DatasetRef(datasetName, database)
-      checkAndAddColumns(dfColumns, dataset, version)
-
-      if (overwrite) {
-        val datasetObj = getDatasetObj(dataset)
-        truncateDataset(datasetObj, version)
-      }
-
-      val numPartitions = df.rdd.partitions.size
-      logger.info(s"Inserting into ($dataset/$version) with $numPartitions partitions")
-
-      // For each partition, start the ingestion
-      df.rdd.mapPartitionsWithIndex { case (index, rowIter) =>
-        // Everything within this function runs on each partition/executor, so need a local datastore & system
-        FiloSetup.init(filoConfig)
-        logger.info(s"Starting ingestion of DataFrame for dataset $dataset, partition $index...")
-        ingestRddRows(FiloSetup.coordinatorActor, dataset, columnNames, version, rowIter,
-                      writeTimeout, index)
-        Iterator.empty
-      }.count()
-
-      // Ensure a flush of memtable after a potentially large ingestion?  But for streaming we might not want
-      // a flush every single time. TODO(velvia): make this configurable
-      if (flushAfterInsert) {
-        actorAsk(FiloSetup.coordinatorActor, Flush(dataset, version)) {
-          case Flushed =>
-          case other: Any => logger.warn(s"Could not finish flushing data!  $other")
-        }
-      }
-
-      syncToHive(sqlContext)
+  private[spark] def dfToFiloColumns(schema: StructType): Seq[DataColumn] = {
+    schema.map { f =>
+      DataColumn(0, f.name, "", -1, sqlTypeToColType(f.dataType))
     }
   }
+
+  private[spark] def checkAndAddColumns(dfColumns: Seq[DataColumn],
+                                        dataset: DatasetRef,
+                                        version: Int): Unit = {
+    // Pull out existing dataset schema
+    val schema = parse(metaStore.getSchema(dataset, version)) { schema =>
+      logger.info(s"Read schema for dataset $dataset = $schema")
+      schema
+    }
+
+    // Translate DF schema to columns, create new ones if needed
+    val dfSchema = dfColumns.map { col => col.name -> col }.toMap
+    val matchingCols = dfSchema.keySet.intersect(schema.keySet)
+    val missingCols = dfSchema.keySet -- schema.keySet
+    logger.info(s"Matching columns - $matchingCols\nMissing columns - $missingCols")
+
+    // Type-check matching columns
+    val matchingTypeErrs = matchingCols.collect {
+      case colName: String if dfSchema(colName).columnType != schema(colName).columnType =>
+        (colName, dfSchema(colName).columnType, schema(colName).columnType)
+    }
+    if (matchingTypeErrs.nonEmpty) throw ColumnTypeMismatch(matchingTypeErrs)
+
+    if (missingCols.nonEmpty) {
+      val addMissingCols = missingCols.map { colName =>
+        val newCol = dfSchema(colName).copy(dataset = dataset.dataset, version = version)
+        metaStore.newColumn(newCol, dataset)
+      }
+      runCommands(addMissingCols)
+    }
+  }
+
+  // Checks for schema errors via RichProjection.make, and returns created Dataset object
+  private[spark] def makeAndVerifyDataset(datasetRef: DatasetRef,
+                                          rowKeys: Seq[String],
+                                          segmentKey: String,
+                                          partitionKeys: Seq[String],
+                                          chunkSize: Option[Int],
+                                          dfColumns: Seq[Column]): Dataset = {
+    val options = Dataset.DefaultOptions
+    val options2 = chunkSize.map { newSize => options.copy(chunkSize = newSize) }.getOrElse(options)
+    val dataset = Dataset(datasetRef, rowKeys, segmentKey, partitionKeys).copy(options = options2)
+
+    // validate against schema.  Checks key names, computed columns, etc.
+    RichProjection.make(dataset, dfColumns).recover {
+      case err: RichProjection.BadSchema => throw BadSchemaError(err.toString)
+    }
+
+    dataset
+  }
+
+  // This doesn't create columns, because that's in checkAndAddColumns.
+  private[spark] def createNewDataset(dataset: Dataset): Unit = {
+    logger.info(s"Creating dataset ${dataset.name}...")
+    actorAsk(FiloSetup.coordinatorActor, CreateDataset(dataset, Nil)) {
+      case DatasetCreated =>
+        logger.info(s"Dataset ${dataset.name} created successfully...")
+      case DatasetError(errMsg) =>
+        throw new RuntimeException(s"Error creating dataset: $errMsg")
+    }
+  }
+
+  private[spark] def truncateDataset(dataset: Dataset, version: Int): Unit = {
+    logger.info(s"Truncating dataset ${dataset.name}")
+    actorAsk(FiloSetup.coordinatorActor,
+             TruncateProjection(dataset.projections.head, version), 1.minute) {
+      case ProjectionTruncated => logger.info(s"Truncation of ${dataset.name} finished")
+      case DatasetError(msg) => throw NotFoundError(s"$msg - (${dataset.name}, ${version})")
+    }
+  }
+
+  private[spark] def deleteDataset(dataset: DatasetRef): Unit = {
+    logger.info(s"Deleting dataset $dataset")
+    parse(FiloSetup.metaStore.deleteDataset(dataset)) { resp => resp }
+  }
+
+  implicit def sqlToFiloContext(sql: SQLContext): FiloContext = new FiloContext(sql)
 }
